[CIVIC-792] Renamed component Label to Field label and updated props, design and refactor other components.
