+++ docroot/themes/contrib/civictheme/civictheme_library/components/01-atoms/field-label/field-label.twig
@@ -1,0 +1,23 @@
+{#
+/**
+ * @file
+ * Label component.
+ *
+ * Variables:
+ * - theme: [string] Theme: light, dark.
+ * - size: [string] Chip size: extra-large, large, regular, small, extra-small.
+ * - content: [string] The label's content.
+ * - attributes: [string] Additional attributes.
+ * - modifier_class: [string] Additional classes.
+ */
+#}
+
+{% set theme_class = 'civictheme-theme-%s'|format(theme|default('light')) %}
+{% set size = size in ['extra-large', 'large', 'regular', 'small', 'extra-small'] ? size : 'regular' %}
+{% set size_class = 'civictheme-field-label--%s'|format(size) %}
+{% set modifier_class = '%s %s %s %s %s %s'|format(theme_class, size_class, modifier_class|default('')) %}
+
+
+{% if content is not empty -%}
+  <label class="civictheme-field-label {{ modifier_class }}" {% if attributes is not empty %}{{ attributes|raw }}{% endif %}>{% if content is not empty %}{{ content }}{% endif %}</label>
+{%- endif %}
