--- docroot/themes/contrib/civictheme/civictheme_library/components/01-atoms/label/label.scss
@@ -1,40 +1,0 @@
-//
-// Label component.
-//
-
-.civictheme-label {
-  $root: &;
-
-  display: block;
-  margin-bottom: civictheme-space();
-
-  @include civictheme-typography('label-regular');
-
-  &--required {
-    &::after {
-      content: ' (Required)';
-
-      @include civictheme-typography('label-small');
-    }
-  }
-
-  &.civictheme-theme-light {
-    color: $civictheme-label-light-color;
-
-    #{$root}--required {
-      &::after {
-        color: $civictheme-label-light-required-color;
-      }
-    }
-  }
-
-  &.civictheme-theme-dark {
-    color: $civictheme-label-dark-color;
-
-    #{$root}--required {
-      &::after {
-        color: $civictheme-label-dark-required-color;
-      }
-    }
-  }
-}
