--- docroot/themes/contrib/civictheme/civictheme_library/components/01-atoms/label/label.twig
@@ -1,24 +1,0 @@
-{#
-/**
- * @file
- * Label component.
- *
- * Variables:
- * - theme: [string] Theme: light, dark.
- * - content: [string] The label's content.
- * - title_display: [string] Elements title_display setting.
- * - required: [boolean] Required or not.
- * - attributes: [string] Additional attributes.
- * - modifier_class: [string] Additional classes.
- */
-#}
-
-{% set theme_class = 'civictheme-theme-%s'|format(theme|default('light')) %}
-{% set title_position = title_display == 'after' ? 'civictheme-label--option' : '' %}
-{% set visibility_class =  title_display == 'invisible' ? 'visually-hidden' : '' %}
-{% set required_class =  required ? 'civictheme-label--required' : '' %}
-{% set modifier_class = '%s %s %s %s %s %s'|format(theme_class, title_position, visibility_class, required_class, required_class, modifier_class|default('')) %}
-
-{% if content is not empty or required -%}
-  <label class="civictheme-label {{ modifier_class }}" {% if attributes is not empty %}{{ attributes|raw }}{% endif %}>{% if content is not empty %}{{ content }}{% endif %}</label>
-{%- endif %}
