+++ docroot/themes/contrib/civictheme/civictheme_library/components/03-organisms/password-indicator/password-indicator.stories.twig
@@ -26,9 +26,8 @@
   'data-password-indicator-setting-min-count="12"',
 ]|join(' ') %}
 {% set username_label %}
-  {% include '@atoms/label/label.twig' with {
+  {% include '@atoms/field-label/field-label.twig' with {
     theme: theme,
-    title_display: 'before',
     content: 'Username',
     attributes: 'for="username"',
   } only %}
@@ -41,9 +40,8 @@
   } only %}
 {% endset %}
 {% set password_label %}
-  {% include '@atoms/label/label.twig' with {
+  {% include '@atoms/field-label/field-label.twig' with {
     theme: theme,
-    title_display: 'before',
     content: 'Password',
     attributes: 'for="password"',
   } only %}
@@ -56,9 +54,8 @@
   } only %}
 {% endset %}
 {% set confirm_label %}
-  {% include '@atoms/label/label.twig' with {
+  {% include '@atoms/field-label/field-label.twig' with {
     theme: theme,
-    title_display: 'before',
     title: 'Confirm Password',
     attributes: 'for="confirm"',
   }%}
